updated STL mesh processing pipeline and cluster filtering
a) replaced current mesh library with mcubes resulting in better formed meshes that can be clustered as meshes before converting to poincloud. this lets us filter out non-candidate meshes and decreases the challenges of clustering pointclouds
b) filtered meshes based on surface area to only incldue things that are the right size for our sphere.